build: add landscape derivation
Adds a Nix derivation for Landscape.

This derivation could stand to be improved, in that it doesn't perform
any management of Node dependencies whatsoever.  It merely copies
existing build artifacts present in the pkg/interface subprojects
directly into the outpath.

This means that when building Landscape, one must, for now, manually run
'make interface' *prior* to running e.g. 'nix-build -A landscape'.